refactor(expr): colocate volatile scalar functions
