+++ src/expr/impl/src/scalar/mod.rs
@@ -103,4 +103,3 @@ mod trim_array;
 mod tumble;
 mod upper;
 mod vector;
-mod volatile;
